refactor: validate() 메서드 내 확인 로직 분리
