+++ src/main/java/lotto/domain/Lotto.java
@@ -1,0 +1,48 @@
+package lotto.domain;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+import static camp.nextstep.edu.missionutils.Randoms.pickUniqueNumbersInRange;
+import static lotto.domain.constant.NumberConstant.*;
+
+public class Lotto {
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        this.numbers = numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        validateSize(numbers);
+        validateDuplicate(numbers);
+    }
+
+    private void validateSize(List<Integer> numbers) {
+        if (numbers.size() != LOTTO_SIZE)
+            throw new IllegalArgumentException();
+    }
+
+    private void validateDuplicate(List<Integer> numbers) {
+        Set<Integer> uniqueNumbers = new HashSet<>();
+
+        for (Integer n : numbers)
+            if (!uniqueNumbers.add(n))
+                throw new IllegalArgumentException();
+    }
+
+
+    public static List<Lotto> createRandomLottos(int ticketNum) {
+        return IntStream.range(0, ticketNum)
+                .mapToObj(i -> generateRandomLotto())
+                .collect(Collectors.toList());
+    }
+
+    private static Lotto generateRandomLotto() {
+        return new Lotto(pickUniqueNumbersInRange(LOTTO_NUMBER_START, LOTTO_NUMBER_END, LOTTO_SIZE));
+    }
+}
